+++ codegen/templates/ethereum/generate.go
@@ -18,7 +18,7 @@ func main() {
 	abi, err := eth.ParseABIFromBytes(abiContent)
 	cli.NoError(err, "Unable to parse ABI file content")
 
-	chain := templates.EthereumChainsByID["ethereum_mainnet"]
+	chain := templates.EthereumChainsByID["Mainnet"]
 
 	project, err := templates.NewEthereumProject("substreams-init-test", "substreams_init_test", chain, eth.MustNewAddress("0xbc4ca0eda7647a8ab7c2061c2e118a18a936f13d"), abi, string(abiContent), 123)
 	cli.NoError(err, "Unable to create Ethereum project")
@@ -26,7 +26,7 @@ func main() {
 	files, err := project.Render()
 	cli.NoError(err, "Unable to render Ethereum project")
 
-	for _, fileToWrite := range []string{"proto/contract.proto", "src/lib.rs", "Cargo.toml", "substreams.yaml"} {
+	for _, fileToWrite := range []string{"proto/contract.proto", "src/lib.rs", "Cargo.toml", "substreams.yaml", "Makefile"} {
 		content, found := files[fileToWrite]
 		cli.Ensure(found, "The file %q is not rendered by Ethereum project", fileToWrite)