+++ codegen/templates/ethereum_project.go
@@ -21,7 +21,7 @@ import (
 //go:embed ethereum/build.rs
 //go:embed ethereum/Cargo.lock
 //go:embed ethereum/Cargo.toml.gotmpl
-//go:embed ethereum/Makefile
+//go:embed ethereum/Makefile.gotmpl
 //go:embed ethereum/substreams.yaml.gotmpl
 //go:embed ethereum/rust-toolchain.toml
 var ethereumProject embed.FS
@@ -66,7 +66,7 @@ func (p *EthereumProject) Render() (map[string][]byte, error) {
 		"build.rs",
 		"Cargo.lock",
 		"Cargo.toml.gotmpl",
-		"Makefile",
+		"Makefile.gotmpl",
 		"substreams.yaml.gotmpl",
 		"rust-toolchain.toml",
 	} {